Added Vi for getting tunnels from while loops as array of clusters.
I had some problems to make it running, but in the end always try to reopen LabVIEW if somethin is not running becuase this is playing mostly with referencies and VI scripting so it's sometimes not running without any reason and LabVIEW restart really helps.
Anyway I created VI which returns list of LoopTunnel for while loops.