Added  indentaion for CaseStructure, and SubVI, make block diagram for transcriptor more readible´, recursive generation of C code for CaseStructure.
I put indentation into generated code and remove some comments, to make it nicer. Also I put there generation of code for each Frame of CaseStructure, so now, when it find CaseStructure, it step through each case and generate code.

I modified this example in LabVIEW 2016 and saved it for 2014 so hope it is still ok.

ToDo:
- recursive generation for diagram
- try to put togethere pieces and parse code for simple LED blinking for example