Repaired recursion, now it's working OK. Corrected VI for getting assignement to outside variables from loops.
After many and many hours looking why is recursion not running I corrected it, but I don't understand today why it's in that way. I hoped that I will get output error terminal from CaseStructure and provide to transcriptor as endpoint for recursion it's inner terminal.
When I did this, recursion doesn't stop, but when I provided wire terminal with index 0 (wire which is connected to the inner terminal of CaseStructure) it's running smoothly.

Anyway I also corrected bug that transcripting ForLoop was crashing , because it wasn't able to count Frames when there wasn't any output from ForLoop, it was because I was counting Frames based on wire names count, but inside terminals on ForLoop doesn't have to be always connected, there even doesn't has to be any output from ForLoop, so now I count Frame count directly as count of InsideTerminal[] what is correct.