Corrected SubVI traversing. Repaired condition to entry recursion. Added example VI for SubVI recursive parsing. Added SubVI password lock detection.
There is no property directly for SubVI which signalize if VI is password locked, so I take SubVI, try to open it and read Diagram, if there is empty constant (there is done just comparison with Constant) it's sign that VI is password locked. Password locked VIs are transcripted by their names to C calls, same for VIs which has in VI name colon (:), they are ssupposed to be part of library (need to distinguisdh btw LabVIEW system library and user defined library).

ToDo:
- checking difference between LabVIEW system VIs and user defined VIs (probably by path)