Repaired code indentation, rework max nested limit as VI, TopLevelDiagram declarations now always moved to top.
I had many mistakes in code indetaion, but now it's done right. Everytime except in case of SubVI when using recursion for code generation, so there hasn't to be defined nested level for recursion, because all children calls are part of superior block and this is indent before added into array.
There is exception in SubVI case, when we are not dealing with password locked VI, so it's going to be translated so next slave calls will be indent to be possible to distinguish them in code (so in case of SubVI it's on us, there is no rule for now).
I put limits for maximal nested level into types/Get Max Nested Level To Go.vi to have it on one place.
TopLevelDiagram is parsed in way that not connected SubVI and not connected String constants are evauluate and wirtten into output. But because constant definition and other stuff is supposed to be in the beggining of each block, I use regular expression and my generated comment to accomplish this. I cut TopLevelDiagram declarations and paste them to beginning of generated block.