Added calling Translator.vi in case of SubVI.
I change transcripting SubVI. Now when there is SubVI, it's checked if in the same directory is file Translator.vi, if yes this Translator.vi is called and the result from it is used.
If there is no Translator.vi it's transcripting in normal way, so using SubVI name.

ToDo:
- writing constant values into TopLevelDiagram wires (it's Variant, so need to create function convert LabVIEW Variant into C value by wire type or something like that)