Removed useless files, there were created at beginning but not used, corrected TopLevelDiagram code output placement.\
Before I had VI which was going through generated code and cut TopLevelDiagram code and put it at beginning, but that was bad concept, because sometinmes there could be som TopLevelDiagram code generated for WhileLoop or ForLoop and this was placed somewhere at generated code at end, so VI was making bad decision and in the end generated bad code.
Now I normaly check array of elements based on their ClassName and if last transcripted element is TopLevelDiagram I just rotate the array with code chunks and after serialiaze to string. That's all what needed.