Added ClusterConstant transcription into datastructure.
ClusterConstant is parsed and for each element is made definition. If it contain another typedef. Cluster, it's used its name, if there is just ClusterConstant which is not typedef, it's traversed and created inside struct as another struct{}.

There is naming rule, if element of Cluster has name, it's used, if no, it's used generated name whi is LabVIEWItemPrefix[index], ie. for 3.item in Cluster wihtout name it will beLabVIEWItem_3 or something else, prefix is stored in types direcotry in VI.