Corrected SubVI to code, if it's not locked transled like normal function and put into definitions. Work on translating clusters.
Clusters would be strcut, and there if are same struct ith same datatypes, they are same, but I have to think about case if there is struct with some user defined value how to assign value if its in struct. I would have to use pointer or some serialization function or it should be defined in C file in header. Think about that. This is work version.