Generating SubVI done, finally generating unlocked SubVI (these wihtout password) and store them as separate functions with params and call them with params using connected wires.
Like I wrote upper, it's all on one place, not divided into subvi, so maybe it's little harder to read, but I tried to put there lot of comments to be clear what I wanted to do. but if there is unlocked SubVI, it's generated it's code inside and stored as separate function in defines.
In code there is just call of SubVI as function by it namse ie. "someSubVINameCall(wireUID_789_, wireUID_111_");", LabVUIEW IDE itself take care no SubVI with same name wuill be loaded, but there would be generated more function with same name because for my transcriptro subvis are translated using wires names and wires UID and they are differen for different copies of same SubVI in memory, but this coul;d be repaire by going through defines and filter them base on function names.

Ok already tired and supposed to go for bed. First I wanted to do this until midnight and now finished at 5am :D Damn, hope at least that it's running somehow good, but Blink3 was able to compile and run without changes.