Corrected "Transcriptors/Transcriptor LabVIEW to C code 2.vi" there was not wired reference terminal for "Get Variant Value.vi" for Cluster reference object.
When "Get Variant Value.vi" is stranslating ClusterConstant it needs reference to this Constant. If it§s not wired there is problem. I corrected it, so now when tyranslating CLusterConstant and has no ref. wire connected putting there just CLuster. It's so because I have to put there some datatype, but don't know which one.
Also I change there in transcripting wires there was owner TopLevelDiagram, this was mistake, because if I put Constant or Bundler or something on Diagram it's owner is ToplevelDiagram until it's not inside some Structure. So rigth code what was supposed to be there was if terminal is owned something what has class ControlTerminal, translate it as Control.