CORRECTED Cluster generation, now it's generating for TopLevelDiagram Cluster definitions if they are or not typedef and after continues normally, if typedef just put there "typedefName variableName;" as normal transcriptor :)
I put into Make Typedef VI getting UID of Terminal to be sure if it's wired. Now it look s it behave all as it has to. For TopLevelDiagram cluster definitions are created and after if tehre is cluster typedef, jus typedef name is used. So simple speaking cluster definition is going for level 1 down and after no. If you want to use your cluster you have to write their definitions into .h file and then include this .h file. Somebody has to try this :) I'm going to try another things.