FIRST NORMAL RUNNING TRANSCRIPTOR. Chappie :)
I repaired wires declaration for while loop to be in front of it with values assign. Also end condition for while loop must be in C code negated becuase in LabVIEW it's condition when to stop while loop, in C code it's condition when program is keeping inside loop.
Now it's able transcript "Blink 4 experimental.vi" into compilable and executable code :)

ToDo:
- some additional function which make this transcriptor usefull (shift registers, ...)
- implement basic arduino api