Added functional feedbaNode implementaion using CircularBuffer from github. Changed directory structure.
Because I changed dir structure I have problem that in translator for system VI is path as string so I had to rewrite it.
I used for feedback node circular registers. There is trick in generating feedback nodes. I had to generate their declaration before structure in which they are, after that I had to define bool value to create single pass case with init values (so var = true, after run case is changed inside to false so it's not running again) and assignements into feedback node has to be at end of diagram declaration, because writing new value into feedback is happening in the end.