Added iterators into loops, changed looking for Translator.vi, now looking for something with Translator in name.
I put there generated iterator for WhileLoop and ForLoop, theirs declarations are created in step 3. Wires are also in step 1. and 2. able assign value from it, this is done by checking if wire is connected to Terminal, then if its owner is Diagram and then if this Diagram owner is Loop (so I now that at this point that wire is connected to iterator terminal).

Because in one project there cannot be two VI with same name, I changed looking for Translator.vi so now its looking for file which has in name Translator word.

Also added folder with translator for my "System API", for now I like to call it this way :D