Corrected getting variant value, now it's handling arrays with max.  dimensions, corrected increment/decrement, now it's correct. Tested on examples "Array 1.vi", there is writing into serial port to see array content.
There was mistake in increment,decrement, there was "%s = %s++" but this is INCORRECT, because it increment variable, but result of this operation is 0 or 1 (successful write), so this is correctly "%s = %s + 1".
Also I repait getting variant value from arrays, it depends if array is 1D, 2D, ..., nD, because there is type cast from variant and at this point I have to know dimensions count, I put there hardocded decision btw. 1D,2D and 3D, arrays with nmore dimensions are not allowed (this will be for Arduino, it hasn't big memory, so 3D array is hopefully enough, otherwise use clusters).
DON'T USE SHIFT REGISTERS FOR 2D ARRAY, IT'S NOT RUNNING. I tried it but program is stucked, instead USE NORMAL TUNNELS for 2D. (for 1D shift registers are fine)