Corrected creating Cluster typedefinitions and datatypes names for wires which carrying Clusters. Also corrected assignin Constant values to wires.
Cluster's datatypes were before generated in way that was whole structure definition created ie. wire was connected to Cluster, so it's datatype was:
struct{int A, String B, vector<String> C} wireUID_420;
It was in that way because if wires are connected to clusters with elements with same datatypes for each of them was created datatyper declaration with same item and we could assign value from one wire to another, but in C this are two different structures and we are getting anonymous structure assignement error, ie.:
struct{int A, String B} wireUID_42 = {17,"Hello world!"};
struct{int A, String B} wireUID_47;
wireUID_47 = wireUID_42;                      <--- THIS IS NOT POSSIBLE, here compiler shows anonymous structure error

So solution to this is create typedefinition and then declares wires using this typedefined structure. So I made it so that I'm generating name for clusters which are not typedefinitions based on their datatypes, so clusters with same datatypes in same order will have generated same datatype name and in this way I can assign values between them:
struct{int A, String B} wireUID_42 = {17,"Hello world!"};       --->       struct_int_String wireUID_42 = {17,"Hello world!"};
struct{int A, String B} wireUID_47;                             --->       struct_int_String wireUID_47;
wireUID_47 = wireUID_42;                                        --->       wireUID_47 = wireUID_42;
So code on right side is valid code and this whole thing now will run.

ToDo:
- I think that know it looks that it runs OK, but there will be probably little confusion, because in step 0. and step 1.,2. I'm using "Cluster Make Typedef.vi" but there are different params used