Directories utils/ and utils2/ merged in utils/
Originaly I was thinking that utils2/ wilkl be extension of utils, but then realize that utils/ are just small functions all should be atomic, so there is no reason to separate them. If I want to make see if some functions are used I can generate class diagram for arduino transcriptor top object and see which ones are used.

Done:
+ I clicked on every VI and use Save as function to move it into utils/ direcotry.
+ I also run transcriptor on each file in Test and mad just one file was no t abe to trasnscript and transcriptor finished with error. So all Test are translatable.
+ also moved Translator.vi and its SubVIs from Test/ to separate directory