Corrected handling class constants on block diagram. Now if wire is connected to class instance it's definition is correctly generated with class datatype, everything else is running same way as before.
For example if there is Servo class constant on diagram wire definition is generated as:
    Servo wireUID_203;
also it's C++ valid definition:
    Servi wireUID_203 = Servo();         (still I'm using first approach)
translating members methods into C++ code is in Translator.vi inside class.

Also there are TranscriptorBasicObject control and indicator placed on transcriptor pallete for user to use them when defines its own library and need to create dynamic dispatch VI.

I had some problems when implementing Servo library, first I had to include basic object into project set it as parent class and then remove it to still be in dependencies and library saved (but still because I installed transcriptor package and also have transcriptor project on other location getting warnings when opening library, but that's just because different location of file on disk, still it's running smoothly for now, need to check and figure out how to make it without any warning).