Made full functional VI for translate code to arduino, possible options in combobox are ESP8266 and Arduino Nano, for now it's just call arduino.exe in shell and put there path to arduino .ino file
Now it's needed to  glu it together, put there path for VI and then translate it by transcriptor, save it as .ino file and compile it for arduino using this VI. Or some similar process, this is the easiest part.

I tested this VI and board spec. is passed into arduino.exe as param, for:
    + esp8266 there is specified in combobox as value esp8266:esp8266:generic:UploadSPeed=...,....
    + for arduino nano there is something like arduino:avr:nano
      ... if you need to spec something else it should be similar, this param is param -fbqn  for arduino-builder.exe ...

OK, i'M SO glad that this is running, tomorrov direction Pohoda 2018 :D :D :D