Corrected transcriptor, there wasn't implemented wathcing reetrant VIs, so in case of recursive loop, transcriptor get stucked in infinite transcription loop. This was solved adding list of reentrant VI into transcriptor engine and disable transcripting whole VI when spoted second time, just their function call.
As written above I forgot about case when VI can be reentrant and I have to remember that I already transcript it's function (I'm talking about non locked VI) and for second time just put C call.
For this I have to add new control to "LabVIEW to C code Engine 2.vi" and for "SubVI Translate.vi", SO THERE IS IN STEP 3 STRING ARRAY PASSING AROUND AND COLLECTING vis REENTRANT NAMES. This array is then used in "SubVI Translate.vi". This has to be wired for every "LabVIEW to C code Engine 2.vi" because we are speaking about language feature, so we have to avoid infinite transcription loop.

NEED TO REBUILD TRANSCRIPTOR PACKAGE.