CORRECTED SYSTEM ERROR, I was putting shift register and feedback nodes wrong, now it's OK, their declarations are in front of loops where they are used, also CASE SELECTOR REPAIRED, it wasn't interpreting enumertion and ring controls right.
For shoft register and feedback nodes I was passing their declarations to their owners, that's wrong, also in code composer there was difference between C code in cluster and direcct C code output, now they are same and shift regs are put on right place.
Case structure selector was wrong, it was interpreting everything as string or number, but enums are represented by string label and they have integer value, now when generating case selector there is check of its datatype and it seems to be correct.

So this release should be most correct release!