Corrected cluster behaviour, now for each cluster even typedef there is created definition and in the end in code composer duplicates are removed. This decision to create definition even for typedef cluster and not to use them as special datatype names (ie. define UserVectorDefCluster and then use it as special datatype, it's in conflict with idea dataflow, so structures are structures, if typedef, its definition has datatype by its name), if somebody wants special datatype use mechanism with classes, you will get even own colored wire.
So clusters are really structures and because they are used in LabVIEW code and they has to be translatable into C++ in the end each cluster definition is placed into code.
Also I added into step 0. scan through all wires and look for cluster and add them into defines list.
Also I improved removing duplicates in the code composer.

This removing duplices and cluster handling is quite dirty after what I rewrite it multiple times, but it's running quite OK for now, it NEEDS REFACTORING at this point (but it's not big deal, code is quite atomized and consistent).

Now this hopefully was last fundamenta system correction.
Enjoy.