Corrected some bugs, includes for nested SubVIs weren't in result C code, that was forgotten copying them into output in step 3 in "utils/SubVI Translate.vi", corrected enter symbol in strings to replaced \n and in flat sequence structure was forgotten to exclude error wires assignements, so it was generated in C code and wires were not defined (error wire is not generated into output it is used to know translation order).
Copying includes for SubVIs is in step 3, there is copying definitions into output but wasn't there copying for includes, now it's OK, so when some nested SubVI was using include like "mySpecialLib.h" this haven't appear in result code. Now it's fixed.
When there was string constant like:
"Hello
      world"
it was as it is in result code because there wasn't replacing enter value for \n, there was replacing newline constant from string pallete to \n but no enter, so I put it there into "Get Variant Value.vi".
For flat sequence in step there is generation of output assignements, I forgot to differ wires datatype and forbid generation for error wire, in the end in code appeared some assignements to this wires, but they were undeclared and compiler showed error. This was resolved by checking wires datatypes and if it's error it's generated empty string which is not passing to output.

Hopefully there is no more this fundamental errors. If yes they will be corrected when appear.

Enjoy.