Corrected type cast, now it should be really doing typecasting, as input could be normal variable or array and output can be also array, classes need to be changed to variant before typecast, for normal retyping is to variant function.
Before it doesn't care about input, there could be normal variable or array and thats make difference, because arrays are transcripted as 'vector<someDatatype>' so to obtain reference to first element we need to use &someVecto.front().
Normal retype:
  A var1;
  B var2;
  var1 = (A) var1;
is done by 'To Variant' function. Classes cannot be typecasted, they have to be first converted to Variant, because Variant is translated as (unsigned int) what is pointer void*.