Added DEBUGGING for MSP430 Launchboard, now there is baseframe for compilation and debugging for MSP430.
Because Energia doesn't suppport vectors (MPS430 is 16-bit MCU, compiler support templates, but STL is not implemented), luckily Code Composer compiler support STL, but in this case is not possible to use Arduino libraries, I put together toolchain to be able compile base code, functions for digitalWrite, analogRead, ... needs to be created (best would be create whole library and create package)
There is in Debuggers/ vi for using debug server which is part of Code Composer. For now this is able to run on my computer and luckily also somewhere else.

Pre-requisities:
  + CCS8, this was created using this code composer
  + CCS installed at c:\ti\ccs8 (this is base path which I used to create this debugger)

What was done:
  + toolchain are copied commands for compiling code from CCS
  + automatized coode compiling and linking (at least for blinking LED)
  + automotized debug of basic code, running debug server and sending request

In Debuggers/ you can find in utils/ some tools copied from CCS8:
  + msp430f5529.ccxml - configuration of MSP430 MCU, needed for compilation and debugging
  + MSP430DebugServer.js - running server app for debugging above TI debug server, listenin to commands and set launchpad board
  + MSP430/utils/MSP430_DebugServer_RUN.bat - windows batch to run, this si specific for my computer, needs to be done in future (paths inside file will be generated automatically)
  + lnk_msp430f5529.cmd - needed for linking to create .out file

THIS VERSION IS EXPERIMENTAL, IT'S ABLE TO RUN ON MY COMPUTER :)

*debug server is not possible shutdown at this moment, it has to be killed in task manager (it appears as ccsstudio and consume on my PC appr. 150MB RAM)